requirements.txt
requests
python-dotenv
gradio
pypdf
google-generativeai